Move state resolving code to the debugger
