+++ src/debugger.ts
@@ -1,4 +1,6 @@
 import type { Scenario } from "./scenario";
+import type { Cache } from "./cache";
+import type { EventID, StateKeyTuple } from "./state_resolver";
 
 // Debugger provides a mechanism for stepping through a scenario, exposing UI elements and calling out to resolve state.
 export class Debugger {
@@ -34,5 +36,57 @@ export class Debugger {
             eventIds.push(this.eventIdOrdering[i]);
         }
         return eventIds;
+    }
+
+    async resolve(
+        cache: Cache,
+        resolveState: (
+            roomId: string,
+            roomVer: string,
+            states: Array<Record<StateKeyTuple, EventID>>,
+        ) => Promise<Record<StateKeyTuple, EventID>>,
+    ): Promise<void> {
+        const atEventId = this.current();
+        const atEvent = cache.eventCache.get(atEventId)!;
+        let theState: Record<StateKeyTuple, EventID> = {};
+        switch (atEvent.prev_events.length) {
+            case 0: // e.g m.room.create
+                // do nothing, as we default to empty prev states.
+                // we'll add the create event now.
+                break;
+            case 1: {
+                // linear: the state is what came before plus this
+                const prevEventId = atEvent.prev_events[0];
+                const prevState = cache.stateAtEvent.getState(prevEventId);
+                if (Object.keys(prevState).length === 0) {
+                    console.error(
+                        `WARN: we do not know the state at ${prevEventId} yet, so the state calculation for ${atEventId} may be wrong!`,
+                    );
+                }
+                theState = prevState;
+                break;
+            }
+            default: {
+                // we need to do state resolution.
+                const states: Array<Record<StateKeyTuple, EventID>> = [];
+                for (const prevEventId of atEvent.prev_events) {
+                    const prevState = cache.stateAtEvent.getState(prevEventId);
+                    if (Object.keys(prevState).length === 0) {
+                        console.error(
+                            `WARN: we do not know the state at ${prevEventId} yet, so the state calculation for ${atEventId} may be wrong!`,
+                        );
+                    }
+                    states.push(prevState);
+                }
+                console.log("performing state resolution for prev_events:", atEvent.prev_events);
+                theState = await resolveState(atEvent.room_id, this.scenario.roomVersion, states);
+            }
+        }
+        // include this state event
+        if (atEvent.state_key != null) {
+            theState[JSON.stringify([atEvent.type, atEvent.state_key])] = atEventId;
+        }
+
+        cache.stateAtEvent.setState(atEventId, theState);
     }
 }
