+++ src/index.ts
@@ -736,56 +736,25 @@ document.getElementById("stepbwd")!.addEventListener("click", async (ev) => {
     dag.refresh();
 });
 document.getElementById("resolve")!.addEventListener("click", async (ev) => {
-    const atEventId = dag.debugger.current();
-    const atEvent = dag.cache.eventCache.get(atEventId)!;
-    let theState: Record<StateKeyTuple, EventID> = {};
-    switch (atEvent.prev_events.length) {
-        case 0: // e.g m.room.create
-            // do nothing, as we default to empty prev states.
-            // we'll add the create event now.
-            break;
-        case 1: {
-            // linear: the state is what came before plus this
-            const prevEventId = atEvent.prev_events[0];
-            const prevState = dag.cache.stateAtEvent.getState(prevEventId);
-            if (Object.keys(prevState).length === 0) {
-                console.error(
-                    `WARN: we do not know the state at ${prevEventId} yet, so the state calculation for ${atEventId} may be wrong!`,
-                );
-            }
-            theState = prevState;
-            break;
-        }
-        default: {
-            // we need to do state resolution.
-            const states: Array<Record<StateKeyTuple, EventID>> = [];
-            for (const prevEventId of atEvent.prev_events) {
-                const prevState = dag.cache.stateAtEvent.getState(prevEventId);
-                if (Object.keys(prevState).length === 0) {
-                    console.error(
-                        `WARN: we do not know the state at ${prevEventId} yet, so the state calculation for ${atEventId} may be wrong!`,
-                    );
-                }
-                states.push(prevState);
-            }
-            console.log("performing state resolution for prev_events:", atEvent.prev_events);
+    await dag.debugger.resolve(
+        dag.cache,
+        async (
+            roomId: string,
+            roomVer: string,
+            states: Array<Record<StateKeyTuple, EventID>>,
+        ): Promise<Record<StateKeyTuple, EventID>> => {
             try {
                 await transport.connect(resolver);
-                const r = await resolver.resolveState(atEvent.room_id, dag.scenario!.roomVersion, states);
-                theState = r.state;
+                const r = await resolver.resolveState(roomId, roomVer, states);
+                return r.state;
             } catch (err) {
                 console.error("failed to setup WS connection:", err);
             } finally {
                 transport.close();
             }
-        }
-    }
-    // include this state event
-    if (atEvent.state_key != null) {
-        theState[JSON.stringify([atEvent.type, atEvent.state_key])] = atEventId;
-    }
-
-    dag.cache.stateAtEvent.setState(atEventId, theState);
+            return {};
+        },
+    );
     dag.refresh();
 });
